Reduce unnecessary copy in main code
`prepend_path` was inputted as a reference, but it was immediately cloned
right after with `PathBuf::from`, even though it was already a `PathBuf`
at the call site in `run_script`.

This commit change the type from a borrowed `&Path` to an owned `PathBuf`,
eliminating the need to re-clone it with `PathBuf::from`.

To @chengcyber, you have the right idea that a borrowed `&Path` is better
than a borrowed `&PathBuf`, but remember, an owned `PathBuf` is a different
story. One must consider how it is used.