Reduce branching
I didn't like how the test of `create_path_env` effectively only test
for one branch of the `if let`: The block under the `if let`. And the
`else` branch was not tested.

So I figured I want to merge the 2 branches into 1. The function
`env::var_os` returns an `Option`, but the `Option` type is an iterator,
I can flat_map it