feat: run script if exists (#47)
pnpm can run scripts even without the `run` subcommand. As in `pnpm run
test` and `pnpm test` being the same. This changes the `Other`
subcommand to check if the first arg is a script in package.json and run
it if exists, otherwise it fallbacks to the previous implementation.

On a side note, the same thing is being done in the `run` subcommand and
the `other` sub, creating duplicate code. Maybe we can move the part of
getting script name from pkg in a func and use it in both subcommand?

I've update the `run_script` test to test both `pn run test` and `pn
test`.

---------

Co-authored-by: khai96_ <hvksmr1996@gmail.com>